+++ CHANGELOG.md
@@ -1,7 +1,21 @@
 # CHANGELOG
 
 **Unpublished:**
-- Introduced CHANGELOG
+- Introduced `CHANGELOG.md`
+- Introduced new services as [social icons](examples/confirm_email.rs):
+  - `Service::Phone` (converts into a link with `tel:` prefix)
+  - `Service::Facebook`
+  - `Service::YouTube`
+  - `Service::Instagram`
+  - `Service::LinkedIn`
+  - `Service::Slack`
+  - `Service::Discord`
+  - `Service::TikTok`
+  - `Service::Snapchat`
+  - `Service::Threads`
+  - `Service::Telegram`
+- **Breaking Changes:**
+  - Renamed `Service::Github` to `Service::GitHub`
 
 **v0.1.0-alpha.4:**
 - Initial implementation